Add a new stats files manipulation class.
The stats module is not sufficiently general and cannot handle the
new call-graph requirements for stats files manipulation.

The module needs a proper rework, however, this would be currently
too time consuming. Hence, this commit adds a new stats files
manipulation class to overcome the current shortcomings.

This class is designed so that it can be easily used in the future
module rework.